Se crean servicios Rest Activos fijos y se ajustan consultas por tipo activo, serial y fecha de compra

+++ src/main/java/com/asd/activos_fijos/entity/ActivosFijos.java
@@ -166,11 +166,11 @@ public class ActivosFijos {
 		this.afFechaCompra = afFechaCompra;
 	}
 
-	public Float getAfValorCompra() {
+	public Double getAfValorCompra() {
 		return afValorCompra;
 	}
 
-	public void setAfValorCompra(Float afValorCompra) {
+	public void setAfValorCompra(Double afValorCompra) {
 		this.afValorCompra = afValorCompra;
 	}
